# api/openapi/v1.yaml
openapi: "3.1.0"
info:
  title: Tofui API
  version: "1.0.0"
  description: OpenTofu Lifecycle Management API
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
paths:
...
  /auth/github/callback:
    get:
      operationId: githubCallback
      summary: GitHub OAuth callback
      tags: [auth]
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "307":
          description: Redirect to frontend with token
...
  /users/{userId}/role:
    put:
      operationId: updateUserRole
      summary: Update user role (owner only)
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [owner, admin, operator, viewer]
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid role or cannot demote last owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /audit-logs:
    get:
      operationId: listAuditLogs
      summary: List audit logs (admin only)
      tags: [audit]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: List of audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        default: 20
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        message:
          type: string
    User:
      type: object
      required: [id, org_id, email, name, role, created_at, updated_at]
      properties:
        id:
          type: string
        org_id:
          type: string
        email:
          type: string
        name:
          type: string
        avatar_url:
          type: string
        role:
          type: string
          enum: [owner, admin, operator, viewer]
        last_login_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AuditLog:
      type: object
      required: [id, org_id, user_id, action, entity_type, entity_id, created_at]
      properties:
        id:
          type: string
        org_id:
          type: string
        user_id:
          type: string
        action:
          type: string
        entity_type:
          type: string
        entity_id:
          type: string
        before_data: {}
        after_data: {}
        ip_address:
          type: string
        user_agent:
          type: string
        created_at:
          type: string
          format: date-time